# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Delivery API",
        "description": "API REST para sistema de gestión de pedidos y entregas",
        "contact": {
            "email": "admin@delivery.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/camiones": {
            "get": {
                "tags": [
                    "Camiones"
                ],
                "summary": "Listar camiones",
                "description": "Obtiene una lista paginada de camiones con búsqueda",
                "operationId": "60ecfe495b428e9c89a1f973da8adc65",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por placa o modelo",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "ABC123"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de camiones obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 15
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Camiones"
                ],
                "summary": "Crear camión",
                "description": "Crea un nuevo camión. Requiere rol admin.",
                "operationId": "e8b3d114ee7a0126e1ee747fe9164e28",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "placa",
                                    "modelo"
                                ],
                                "properties": {
                                    "placa": {
                                        "type": "string",
                                        "maxLength": 10,
                                        "minLength": 5,
                                        "example": "ABC123"
                                    },
                                    "modelo": {
                                        "type": "string",
                                        "maxLength": 10,
                                        "minLength": 2,
                                        "example": "2020"
                                    },
                                    "camioneros": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Camión creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado - Se requiere rol admin"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/camioneros": {
            "get": {
                "tags": [
                    "Camioneros"
                ],
                "summary": "Listar camioneros",
                "description": "Obtiene una lista paginada de camioneros con búsqueda",
                "operationId": "667ed993a34fbfc189dab0187f6665ac",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por nombre, apellido, documento o licencia",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Juan"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de camioneros obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 25
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Camioneros"
                ],
                "summary": "Crear camionero",
                "description": "Crea un nuevo camionero. Requiere rol admin. El camionero debe ser mayor de 18 años.",
                "operationId": "004dbdfd62bac89338b1c0adf96581f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "documento",
                                    "nombre",
                                    "apellido",
                                    "fecha_nacimiento",
                                    "licencia",
                                    "telefono"
                                ],
                                "properties": {
                                    "documento": {
                                        "type": "string",
                                        "maxLength": 10,
                                        "minLength": 5,
                                        "example": "1234567890"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "maxLength": 45,
                                        "minLength": 2,
                                        "example": "Juan"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "maxLength": 45,
                                        "minLength": 2,
                                        "example": "Pérez"
                                    },
                                    "fecha_nacimiento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-15"
                                    },
                                    "licencia": {
                                        "type": "string",
                                        "maxLength": 10,
                                        "minLength": 5,
                                        "example": "ABC12345"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "maxLength": 15,
                                        "minLength": 7,
                                        "example": "+57 300 1234567"
                                    },
                                    "camiones": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Camionero creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado - Se requiere rol admin"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}